Expose phsysical x2APIC IDs when needed
Previously, we exposed the x2APIC IDs at the boot time. It
required us to provide the fixed VCPU-PCPU mapping beforehand
and thus, reduced the degree of load balancing among PCPUs. We
eleviated such requirements and exposed the specified x2APIC
IDs when needed. It would be the responsibility of admin or
hypervisor to determine which PCPUs were available for the
bare-metal virtualization.

modified:   x86-kvm/lapic.c
modified:   x86-kvm/x86.c